tests: fs: Make mount flags test common
Expose test_fs_mount_flags test for other file systems' tests.
