+++ tests/subsys/fs/littlefs/CMakeLists.txt
@@ -15,4 +15,8 @@ target_compile_definitions(app PRIVATE
 target_sources(app PRIVATE
   ${app_sources}
   ../common/test_fs_open_flags.c
-  ../common/test_fs_mkfs.c)
+  ../common/test_fs_dirops.c
+  ../common/test_fs_basic.c
+  ../common/test_fs_mount_flags.c
+  ../common/test_fs_mkfs.c
+)
